feat(financeiro a1): refator FinanceiroCapitalGiro em 4 tabs (eventos plugado, resto placeholder)
